Fixes FOUC on navbar icons.
This is a "Flash Of Unstyled Content" bug that happens with FontAwesome's icons. While the initial JS is loading, the icons are rendered at their inherent size (the SVG's are based on a 512px grid, so it uses that).

This commit adds styling to the component in a way that bypasses the React component lifecycle, so it doesn't require JS to be loaded and will instead be included with the initial page payload. Yay!